requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pydantic>=1.9.0
email-validator>=1.1.3
python-multipart>=0.0.5
python-dotenv>=0.19.0
flake8>=6.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
scikit-learn>=1.0.2
numpy>=1.22.0